Use virtualenv from st2 virtualenv by default
If a virtualenv was built with symlinks instead of copies of the python
binary, as happens by default on GHA integration tests, then we end up
using the system virtualenv binary instead of what we installed.

Currently GHA has the same version of virtualenv, but we build
virtualenvs with --system-site-packages, and we need it to use the st2
virtualenv as the "system" dir instead of the GHA provided one. So,
don't get the realpath of the python binary.