Import PATH/NAME for each fixture pack part 2
In order for pants to detect which changes need to be re-run after changing
fixtures, pants needs to be able to infer the relevant dependencies.

We could use a super-level target to depend on everything, but then we lose
the fine-grained invalidation. The only way to tell pants that a test file
is using a particular set of fixtures is via python imports.

So, this adds some boilerplate python in every pack dir:
__init__.py
fixture.py